move and improve archive spec
